requirements.txt
Flask>=3.0.0
Pillow>=10.0.0
flask-cors>=4.0.0
google-genai>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.7.0
fastapi>=0.115.0
uvicorn>=0.30.0
